moved pure functions to ext.rs
