+++ src/ext.rs
@@ -122,9 +122,9 @@ pub fn collapse(state: &[Complex<f64>]) -> usize {
     dist.sample(&mut rng)
 }
 
-/// # measure
+/// # measure_and_observe_sv
 /// Returns a probable state vector after measurement.
-pub fn measure(
+pub fn measure_and_observe_sv(
     target: usize,
     state: &DVector<Complex<f64>>,
     n_qubits: usize,
@@ -263,6 +263,105 @@ pub fn expand_matrix(
             }
         }
         sum += eval_tensor_product(term);
+    }
+    sum
+}
+
+/// # measure_and_observe_dm
+/// Returns the mesurement a probable density matrix after measurement.
+pub fn measure_and_observe_dm(
+    density: &DMatrix<Complex<f64>>,
+    target: usize,
+    n_qubits: usize,
+) -> (usize, DMatrix<Complex<f64>>) {
+    // Choose a collapsed state
+    let prob_target_eq_zero = density
+        .diagonal()
+        .iter()
+        .enumerate()
+        .filter(|&(idx, _)| (1 << target) & idx == 0) // Using |..q_1q_0> convetion
+        .map(|(_, c)| c.re) // Diagonal elements of density matrices should always be positive & real.
+        .sum::<f64>();
+
+    let mut rng = rand::rng();
+    let random_value = rng.random_range(0.0..1.0);
+    let mut result = 1;
+    let mut result_density = dmatrix![cart!(0.0), cart!(0.0); cart!(0.0), cart!(1.0)]; // |1><1|
+    if random_value < prob_target_eq_zero {
+        // 0 was chosen as collapsed state.
+        result = 0;
+        result_density = dmatrix![cart!(1.0), cart!(0.0); cart!(0.0), cart!(0.0)]; // |0><0|
+    }
+
+    // Calculate projection operator, M
+    let mut projection_operator_prod = identity_tensor_factors(n_qubits);
+    projection_operator_prod[target] = result_density;
+    let projection_operator = eval_tensor_product(projection_operator_prod);
+
+    /* Use formula for next state:
+     *
+     *               MpM'
+     *  p`  ==   ___________
+     *            tr(M'Mp)
+     *
+     * Note that M is always diagonal and has only elements 0 and 1. Such matricies
+     * must be Hermitian so: M' == M. and we can rewrite as:
+     *
+     *               MpM
+     *  p`  ==   ___________
+     *            tr(M^2p)
+     *
+     * Note that M always being diagonal and elements being 0 and 1 ==> M^2 == M.
+     * We can therefore rewrite the formula as:
+     *
+     *               MpM
+     *  p`  ==   ___________
+     *             tr(Mp)
+     *
+     * */
+
+    let proj_op_times_density = projection_operator.clone() * density; //Mp
+    let norm = proj_op_times_density.trace(); // tr(Mp)
+    (result, proj_op_times_density * projection_operator / norm)
+}
+
+/// # reduced_state
+/// Returns the reduced state (partial trace) of a density matrix,
+/// where `targets` specifies the subsystem of qubits.
+pub fn reduced_state(
+    density: &DMatrix<Complex<f64>>,
+    targets: &[usize],
+    n_qubits: usize,
+) -> DMatrix<Complex<f64>> {
+    let bra = [
+        dmatrix![cart!(1.0), cart!(0.0)], // <0|
+        dmatrix![cart!(0.0), cart!(1.0)], // <1|
+    ];
+
+    let dim = 1 << targets.len();
+    let mut sum = DMatrix::<Complex<f64>>::zeros(dim, dim);
+
+    let non_targets = (0..n_qubits)
+        .filter(|idx| !targets.contains(idx))
+        .collect::<Vec<usize>>();
+    let n_terms = 1 << (n_qubits - targets.len());
+    for i in 0..n_terms {
+        /* Example, targets = [2], n_qubits = 3:
+         *
+         * p_2 = (<0| * <0| * I)p(|0> * |0> * I) +
+         *     + (<0| * <1| * I)p(|0> * |1> * I) +
+         *     + (<1| * <0| * I)p(|1> * |0> * I) +
+         *     + (<1| * <1| * I)p(|1> * |1> * I) +
+         * */
+        let mut left_of_density_prod = identity_tensor_factors(n_qubits);
+        let mut j: usize = 0;
+        for non_target in non_targets.iter() {
+            left_of_density_prod[*non_target] = bra[(i >> j) & 1].clone();
+            j += 1;
+        }
+        let left_of_density = eval_tensor_product(left_of_density_prod);
+        let right_of_density = left_of_density.adjoint();
+        sum += left_of_density * density * right_of_density;
     }
     sum
 }
